requirements.txt
# GPU Orchestration – Python dependencies (production pinned)
# Install: pip install -r requirements.txt

# ── Web framework ────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# ── HTTP client ──────────────────────────────────────────────
httpx>=0.27.0
requests>=2.31.0

# ── Database ORM ─────────────────────────────────────────────
sqlalchemy>=2.0.30
alembic>=1.13.0
# Uncomment for PostgreSQL:
# psycopg2-binary>=2.9.9

# ── Configuration & secrets ──────────────────────────────────
pyyaml>=6.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# ── Security ─────────────────────────────────────────────────
# (API key auth is built-in; add these if you extend auth later)
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4

# ── Observability ────────────────────────────────────────────
# Uncomment to enable Prometheus /metrics endpoint:
# prometheus-fastapi-instrumentator>=7.0.0

# ── Deep learning ─────────────────────────────────────────────
# CPU-only (default). For CUDA, replace with:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0

# ── System metrics ───────────────────────────────────────────
psutil>=5.9.8
